backend/requirement.txt
fastapi
uvicorn
python-dotenv
google-generativeai
PyPDF2
python-multipart